+++ services/firebase.mjs
@@ -1,48 +1,52 @@
 import { db } from "./database.mjs";
 import { collection, doc, addDoc, getDocs, setDoc } from "firebase/firestore"; 
+import { getRandomDate, generateRandomBoolean } from "./utility.mjs";
 
+const APPLICATIONS = "applications";
+const APPLICANTS = "applicants";
 
-export async function test_database(){
-    try {
-        const doc_id = String(Math.floor(Math.random()*100000000000));
-        const applicant_id = String(Math.floor(Math.random()*100000000000));
-        const steps = Math.floor(Math.random()* 3);
-        const application_ref = collection(db,"application_test")
+export async function createApplication(application_id, applicant_id){
+    const steps = Math.floor(Math.random()* 3);
+    const application_ref = collection(db,APPLICATIONS);
 
-        const docRef = await setDoc(doc(application_ref, doc_id), {
-        UID: doc_id,
+    return await setDoc(doc(application_ref, application_id), {
+        UID: application_id,
         applicant_ID: applicant_id,
         step: steps,
-        webinar_date: new Date(),
-        approved: false,
-        docs_provided: false,
-        is_complete: false
+        webinar_date: getRandomDate(30),
+        approved: generateRandomBoolean(),
+        docs_provided: generateRandomBoolean(),
+        is_complete: generateRandomBoolean()
     });
-        console.log("Applicant created");
-    } catch (e) {
-        console.error("Error adding document: ", e);
-    }
+   
     }
     
-
-export async function fetchApplicants(){
+export async function fetchCollection(collection_name){
     try{
-        const querySnapshot = await getDocs(collection(db, 'applicants'));
-        querySnapshot.forEach((i)=> {
-        console.log("data: ", i.data());
-    })
+        const querySnapshot = await getDocs(collection(db, collection_name));
+        const documents = []
+        querySnapshot.forEach((doc)=> {
+            documents.push(doc.data());
+        })
+        return documents;
     }
     catch(e){
-        console.log(e)
+        return []
     }
 } 
+export async function fetchApplicants(){
+    return await fetchCollection(APPLICANTS)
+} 
 
-export async function create_applicants_firebase(applicant_list){
+export async function fetchApplications(){
+    return await fetchCollection(APPLICATIONS)
+}
+
+export async function createApplicantsFirebase(applicant_list){
     try {
         const promise_collection = []
         applicant_list.forEach(async (applicant)=>{
-            console.log(applicant["UID"])
-            const collection_ref = collection(db,"applicants");
+            const collection_ref = collection(db,APPLICANTS);
             const doc_ref = doc(collection_ref, applicant["UID"]);
             promise_collection.push(await setDoc(doc_ref, applicant));
         })
@@ -52,7 +56,6 @@ export async function create_applicants_firebase(applicant_list){
         }).catch((e)=>{
            console.log("failed");
              return String(e);
-           
         })
     }
     catch(e){
@@ -60,19 +63,19 @@ export async function create_applicants_firebase(applicant_list){
         console.log(e);
         return e
     }
-
 }
+
 export async function SetApproval(isTrue, UID){
-    const applicationRef = doc(db, 'application_test', UID);
-    setDoc(applicationRef, { approved: isTrue}, { merge: true } );
+    const applicationRef = doc(db, APPLICATIONS, UID);
+    return await setDoc(applicationRef, { approved: isTrue}, { merge: true } );
 }
 
 export async function SetProvidedDocumentation(isTrue, UID){
-    const applicationRef = doc(db, 'application_test', UID);
-    setDoc(applicationRef, { docs_provided: isTrue}, { merge: true } );
+    const applicationRef = doc(db, APPLICATIONS, UID);
+    return await setDoc(applicationRef, { docs_provided: isTrue}, { merge: true } );
 }
 
 export async function SetStep(ID, stepNum){
-    const applicationRef = doc(db, 'application_test', ID);
-    setDoc(applicationRef, { step: stepNum}, { merge: true } );
+    const applicationRef = doc(db, APPLICATIONS, ID);
+    return await setDoc(applicationRef, { step: stepNum}, { merge: true } );
 }
